requirements_311.txt
bokeh>=3.5.2
hdbscan>=0.8.38.post1
littleballoffur>=2.3.1
matplotlib>=3.9.2
networkx>=3.3
numpy>=1.24.0
ogb>=1.3.6
pandas>=1.5.3
PyYAML>=6.0.2
PyYAML>=6.0.2
rdkit>=2024.3.5
scikit_learn>=1.5.2
seaborn>=0.13.2
# torch>=2.4
torch_geometric # >=2.6.0
torch_scatter # >=2.1.2
torch_sparse # >=0.6.18
tqdm>=4.66.5
umap_learn>=0.5.6
wget>=3.2
